TST: Move groupby tests out of test_function
